+++ aws-lambda-router/CHANGELOG.md
@@ -4,6 +4,10 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.38.1](https://github.com/wundergraph/cosmo/compare/aws-lambda-router@0.38.0...aws-lambda-router@0.38.1) (2025-04-02)
+
+**Note:** Version bump only for package aws-lambda-router
+
 # [0.38.0](https://github.com/wundergraph/cosmo/compare/aws-lambda-router@0.37.7...aws-lambda-router@0.38.0) (2024-12-21)
 
 ### Features