+++ controlplane/CHANGELOG.md
@@ -4,6 +4,16 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.132.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.131.1...controlplane@0.132.0) (2025-04-04)
+
+### Features
+
+* improve delete organization workflow ([#1741](https://github.com/wundergraph/cosmo/issues/1741)) ([643c179](https://github.com/wundergraph/cosmo/commit/643c179f71daf5a77a56c685b3924d10894c2d9c)) (@wilsonrivera)
+
+## [0.131.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.131.0...controlplane@0.131.1) (2025-04-02)
+
+**Note:** Version bump only for package controlplane
+
 # [0.131.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.130.0...controlplane@0.131.0) (2025-04-01)
 
 ### Features